Update and add Build Action
Update and add the Build Action to build bases.

This implmenets the following
* Build Action
* Docker Layer Cache
  * Argument to specify docker layer cache within GH action cache
* Matrix output to support matrix strategy for building the image
* DockerFile cache Key generator (create a cache key based upon the
  docker file SHA256)

This is all in support of consistent and reproducible builds. We may
need to think about the docker file content cache as it may make
security updates more challenging.

Image push is currently a stub and does not actually push images